cyber-triage-tool/requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
psutil>=5.9.0
requests>=2.28.0
flask>=2.3.0
flask-socketio>=5.3.0

# Security and forensics libraries
yara-python>=4.3.0
python-magic>=0.4.27
cryptography>=41.0.0
pefile>=2023.2.7
volatility3>=2.5.0

# Network analysis
scapy>=2.5.0
netaddr>=0.8.0
dnspython>=2.3.0

# File analysis
oletools>=0.60
python-evtx>=0.7.4
regipy>=3.0.0

# Database
sqlalchemy>=2.0.0

# Web interface
jinja2>=3.1.0
bootstrap-flask>=2.2.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
rich>=13.0.0
colorama>=0.4.6
tqdm>=4.65.0
python-dateutil>=2.8.0
pathlib2>=2.3.7

# Testing (dev dependencies)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional ML/AI for advanced analysis
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
